Getting close now.

+++ sipsorcery-silverlight/SIPSorcery.GUI/Services/SIPSorceryWebServicePersistor.cs
@@ -11,14 +11,13 @@ using System.Windows.Input;
 using System.Windows.Media;
 using System.Windows.Media.Animation;
 using System.Windows.Shapes;
+using SIPSorcery.CRM;
 using SIPSorcery.SIP.App;
 using SIPSorcery.Silverlight.Messaging;
 using SIPSorcery.SIPSorceryProvisioningClient;
 
-namespace SIPSorcery.Persistence
-{
-    public class SIPSorceryWebServicePersistor : SIPSorceryPersistor
-    {
+namespace SIPSorcery.Persistence {
+    public class SIPSorceryWebServicePersistor : SIPSorceryPersistor {
         public override event IsAliveCompleteDelegate IsAliveComplete;
         public override event LoginCompleteDelegate LoginComplete;
         public override event LogoutCompleteDelegate LogoutComplete;
@@ -46,11 +45,14 @@ namespace SIPSorcery.Persistence
         public override event GetCallsCompleteDelegate GetCallsComplete;
         public override event GetCDRsCountCompleteDelegate GetCDRsCountComplete;
         public override event GetCDRsCompleteDelegate GetCDRsComplete;
+        public override event CreateCustomerCompleteDelegate CreateCustomerComplete;
+        public override event DeleteCustomerCompleteDelegate DeleteCustomerComplete;
+
+        public override event MethodInvokerDelegate SessionExpired = () => { };
 
         private SIPProvisioningWebServiceClient m_provisioningServiceProxy;
 
-        public SIPSorceryWebServicePersistor(string serverURL, string authid)
-        {
+        public SIPSorceryWebServicePersistor(string serverURL, string authid) {
             //BasicHttpBinding binding = new BasicHttpBinding();
             BasicHttpCustomHeaderBinding binding = new BasicHttpCustomHeaderBinding(new SecurityHeader(authid));
             EndpointAddress address = new EndpointAddress(serverURL);
@@ -84,15 +86,31 @@ namespace SIPSorcery.Persistence
             m_provisioningServiceProxy.GetCallsCompleted += m_provisioningServiceProxy_GetCallsCompleted;
             m_provisioningServiceProxy.GetCDRsCountCompleted += GetCDRsCountCompleted;
             m_provisioningServiceProxy.GetCDRsCompleted += GetCDRsCompleted;
+            m_provisioningServiceProxy.CreateCustomerCompleted += CreateCustomerCompleted;
+            m_provisioningServiceProxy.DeleteCustomerCompleted += DeleteCustomerCompleted;
         }
 
-        public override void IsAliveAsync()
-        {
+        private bool IsUnauthorised(Exception excp) {
+            if (excp != null && excp.GetType() == typeof(RawFaultException)) {
+                RawFaultException rawExcp = (RawFaultException)excp;
+                if (rawExcp.FaultType == typeof(UnauthorizedAccessException)) {
+                    return true;
+                }
+            }
+            return false;
+        }
+
+        public override void IsAliveAsync() {
             m_provisioningServiceProxy.IsAliveAsync();
         }
 
         private void IsAliveCompleted(object sender, IsAliveCompletedEventArgs e) {
-            IsAliveComplete(e);
+            if (IsUnauthorised(e.Error)) {
+                SessionExpired();
+            }
+            else {
+                IsAliveComplete(e);
+            }
         }
 
         public override void LoginAsync(string username, string password) {
@@ -108,11 +126,16 @@ namespace SIPSorcery.Persistence
         }
 
         private void LogoutCompleted(object sender, AsyncCompletedEventArgs e) {
-                LogoutComplete(e);
+            LogoutComplete(e);
         }
 
         private void GetSIPDomainsCompleted(object sender, GetSIPDomainsCompletedEventArgs e) {
-            GetSIPDomainsComplete(e);
+            if (IsUnauthorised(e.Error)) {
+                SessionExpired();
+            }
+            else {
+                GetSIPDomainsComplete(e);
+            }
         }
 
         #region SIP Accounts.
@@ -122,7 +145,12 @@ namespace SIPSorcery.Persistence
         }
 
         private void GetSIPAccountsCountCompleted(object sender, GetSIPAccountsCountCompletedEventArgs e) {
-            GetSIPAccountsCountComplete(e);
+            if (IsUnauthorised(e.Error)) {
+                SessionExpired();
+            }
+            else {
+                GetSIPAccountsCountComplete(e);
+            }
         }
 
         public override void GetSIPAccountsAsync(string whereExpression, int offset, int count) {
@@ -130,7 +158,12 @@ namespace SIPSorcery.Persistence
         }
 
         private void GetSIPAccountsCompleted(object sender, GetSIPAccountsCompletedEventArgs e) {
-            GetSIPAccountsComplete(e);
+            if (IsUnauthorised(e.Error)) {
+                SessionExpired();
+            }
+            else {
+                GetSIPAccountsComplete(e);
+            }
         }
 
         public override void AddSIPAccountAsync(SIPAccount sipAccount) {
@@ -138,7 +171,12 @@ namespace SIPSorcery.Persistence
         }
 
         private void AddSIPAccountCompleted(object sender, AddSIPAccountCompletedEventArgs e) {
-            AddSIPAccountComplete(e);
+            if (IsUnauthorised(e.Error)) {
+                SessionExpired();
+            }
+            else {
+                AddSIPAccountComplete(e);
+            }
         }
 
         public override void UpdateSIPAccount(SIPAccount sipAccount) {
@@ -146,7 +184,12 @@ namespace SIPSorcery.Persistence
         }
 
         private void UpdateSIPAccountCompleted(object sender, UpdateSIPAccountCompletedEventArgs e) {
-            UpdateSIPAccountComplete(e);
+            if (IsUnauthorised(e.Error)) {
+                SessionExpired();
+            }
+            else {
+                UpdateSIPAccountComplete(e);
+            }
         }
 
         public override void DeleteSIPAccount(SIPAccount sipAccount) {
@@ -154,7 +197,12 @@ namespace SIPSorcery.Persistence
         }
 
         private void DeleteSIPAccountCompleted(object sender, DeleteSIPAccountCompletedEventArgs e) {
-            DeleteSIPAccountComplete(e);
+            if (IsUnauthorised(e.Error)) {
+                SessionExpired();
+            }
+            else {
+                DeleteSIPAccountComplete(e);
+            }
         }
 
         #endregion
@@ -166,7 +214,12 @@ namespace SIPSorcery.Persistence
         }
 
         private void GetSIPProvidersCountCompleted(object sender, GetSIPProvidersCountCompletedEventArgs e) {
-            GetSIPProvidersCountComplete(e);
+            if (IsUnauthorised(e.Error)) {
+                SessionExpired();
+            }
+            else {
+                GetSIPProvidersCountComplete(e);
+            }
         }
 
         public override void GetSIPProvidersAsync(string where, int offset, int count) {
@@ -174,7 +227,12 @@ namespace SIPSorcery.Persistence
         }
 
         private void GetSIPProvidersCompleted(object sender, GetSIPProvidersCompletedEventArgs e) {
-            GetSIPProvidersComplete(e);
+            if (IsUnauthorised(e.Error)) {
+                SessionExpired();
+            }
+            else {
+                GetSIPProvidersComplete(e);
+            }
         }
 
         public override void DeleteSIPProviderAsync(SIPProvider sipProvider) {
@@ -182,7 +240,12 @@ namespace SIPSorcery.Persistence
         }
 
         private void DeleteSIPProviderCompleted(object sender, DeleteSIPProviderCompletedEventArgs e) {
-            DeleteSIPProviderComplete(e);
+            if (IsUnauthorised(e.Error)) {
+                SessionExpired();
+            }
+            else {
+                DeleteSIPProviderComplete(e);
+            }
         }
 
         public override void UpdateSIPProviderAsync(SIPProvider sipProvider) {
@@ -190,7 +253,12 @@ namespace SIPSorcery.Persistence
         }
 
         private void UpdateSIPProviderCompleted(object sender, UpdateSIPProviderCompletedEventArgs e) {
-            UpdateSIPProviderComplete(e);
+            if (IsUnauthorised(e.Error)) {
+                SessionExpired();
+            }
+            else {
+                UpdateSIPProviderComplete(e);
+            }
         }
 
         public override void AddSIPProviderAsync(SIPProvider sipProvider) {
@@ -198,7 +266,12 @@ namespace SIPSorcery.Persistence
         }
 
         private void AddSIPProviderCompleted(object sender, AddSIPProviderCompletedEventArgs e) {
-            AddSIPProviderComplete(e);
+            if (IsUnauthorised(e.Error)) {
+                SessionExpired();
+            }
+            else {
+                AddSIPProviderComplete(e);
+            }
         }
 
         public override void GetSIPProviderBindingsAsync(string where, int offset, int count) {
@@ -206,7 +279,12 @@ namespace SIPSorcery.Persistence
         }
 
         private void GetSIPProviderBindingsCompleted(object sender, GetSIPProviderBindingsCompletedEventArgs e) {
-            GetSIPProviderBindingsComplete(e);
+            if (IsUnauthorised(e.Error)) {
+                SessionExpired();
+            }
+            else {
+                GetSIPProviderBindingsComplete(e);
+            }
         }
 
         public override void GetSIPProviderBindingsCountAsync(string whereExpression) {
@@ -214,74 +292,79 @@ namespace SIPSorcery.Persistence
         }
 
         private void GetSIPProviderBindingsCountCompleted(object sender, GetSIPProviderBindingsCountCompletedEventArgs e) {
-            GetSIPProviderBindingsCountComplete(e);
+            if (IsUnauthorised(e.Error)) {
+                SessionExpired();
+            }
+            else {
+                GetSIPProviderBindingsCountComplete(e);
+            }
         }
 
         #endregion
 
         #region Dial Plans.
 
-        private void GetDialPlansCountCompleted(object sender, GetDialPlansCountCompletedEventArgs e)
-        {
-            if (GetDialPlansCountComplete != null)
-            {
+        private void GetDialPlansCountCompleted(object sender, GetDialPlansCountCompletedEventArgs e) {
+            if (IsUnauthorised(e.Error)) {
+                SessionExpired();
+            }
+            else {
                 GetDialPlansCountComplete(e);
             }
         }
 
-        public override void GetDialPlansCountAsync(string where)
-        {
+        public override void GetDialPlansCountAsync(string where) {
             m_provisioningServiceProxy.GetDialPlansCountAsync(where);
         }
 
-        private void GetDialPlansCompleted(object sender, GetDialPlansCompletedEventArgs e)
-        {
-            if (GetDialPlansComplete != null)
-            {
+        private void GetDialPlansCompleted(object sender, GetDialPlansCompletedEventArgs e) {
+            if (IsUnauthorised(e.Error)) {
+                SessionExpired();
+            }
+            else {
                 GetDialPlansComplete(e);
             }
         }
-        
-        public override void GetDialPlansAsync(string where, int offset, int count)
-        {
+
+        public override void GetDialPlansAsync(string where, int offset, int count) {
             m_provisioningServiceProxy.GetDialPlansAsync(where, offset, count);
         }
 
-        public override void DeleteDialPlanAsync(SIPDialPlan dialPlan)
-        {
+        public override void DeleteDialPlanAsync(SIPDialPlan dialPlan) {
             m_provisioningServiceProxy.DeleteDialPlanAsync(dialPlan);
         }
 
-        private void DeleteDialPlanCompleted(object sender, DeleteDialPlanCompletedEventArgs e)
-        {
-            if (DeleteDialPlanComplete != null)
-            {
+        private void DeleteDialPlanCompleted(object sender, DeleteDialPlanCompletedEventArgs e) {
+            if (IsUnauthorised(e.Error)) {
+                SessionExpired();
+            }
+            else {
                 DeleteDialPlanComplete(e);
             }
         }
 
-        public override void AddDialPlanAsync(SIPDialPlan dialPlan)
-        {
+        public override void AddDialPlanAsync(SIPDialPlan dialPlan) {
             m_provisioningServiceProxy.AddDialPlanAsync(dialPlan);
         }
 
-        private void AddDialPlanCompleted(object sender, AddDialPlanCompletedEventArgs e)
-        {
-            if (AddDialPlanComplete != null)
-            {
+        private void AddDialPlanCompleted(object sender, AddDialPlanCompletedEventArgs e) {
+            if (IsUnauthorised(e.Error)) {
+                SessionExpired();
+            }
+            else {
                 AddDialPlanComplete(e);
             }
         }
 
-        public override void UpdateDialPlanAsync(SIPDialPlan dialPlan)
-        {
+        public override void UpdateDialPlanAsync(SIPDialPlan dialPlan) {
             m_provisioningServiceProxy.UpdateDialPlanAsync(dialPlan);
         }
 
-        private void UpdateDialPlanCompleted(object sender, UpdateDialPlanCompletedEventArgs e)
-        {
-            if (UpdateDialPlanComplete != null)
-            {
+        private void UpdateDialPlanCompleted(object sender, UpdateDialPlanCompletedEventArgs e) {
+            if (IsUnauthorised(e.Error)) {
+                SessionExpired();
+            }
+            else {
                 UpdateDialPlanComplete(e);
             };
         }
@@ -297,7 +380,12 @@ namespace SIPSorcery.Persistence
         }
 
         private void GetSIPRegistrarBindingsCompleted(object sender, GetSIPRegistrarBindingsCompletedEventArgs e) {
-            GetRegistrarBindingsComplete(e);
+            if (IsUnauthorised(e.Error)) {
+                SessionExpired();
+            }
+            else {
+                GetRegistrarBindingsComplete(e);
+            }
         }
 
         public override void GetRegistrarBindingsCountAsync(string whereExpression) {
@@ -305,7 +393,12 @@ namespace SIPSorcery.Persistence
         }
 
         private void GetSIPRegistrarBindingsCountCompleted(object sender, GetSIPRegistrarBindingsCountCompletedEventArgs e) {
-            GetRegistrarBindingsCountComplete(e);
+            if (IsUnauthorised(e.Error)) {
+                SessionExpired();
+            }
+            else {
+                GetRegistrarBindingsCountComplete(e);
+            }
         }
 
         public override void GetCallsCountAsync(string whereExpression) {
@@ -313,7 +406,12 @@ namespace SIPSorcery.Persistence
         }
 
         private void m_provisioningServiceProxy_GetCallsCountCompleted(object sender, GetCallsCountCompletedEventArgs e) {
-            GetCallsCountComplete(e);
+            if (IsUnauthorised(e.Error)) {
+                SessionExpired();
+            }
+            else {
+                GetCallsCountComplete(e);
+            }
         }
 
         public override void GetCallsAsync(string whereExpressionn, int offset, int count) {
@@ -321,7 +419,12 @@ namespace SIPSorcery.Persistence
         }
 
         private void m_provisioningServiceProxy_GetCallsCompleted(object sender, GetCallsCompletedEventArgs e) {
-            GetCallsComplete(e);
+            if (IsUnauthorised(e.Error)) {
+                SessionExpired();
+            }
+            else {
+                GetCallsComplete(e);
+            }
         }
 
         public override void GetCDRsCountAsync(string whereExpression) {
@@ -329,7 +432,12 @@ namespace SIPSorcery.Persistence
         }
 
         private void GetCDRsCountCompleted(object sender, GetCDRsCountCompletedEventArgs e) {
-            GetCDRsCountComplete(e);
+            if (IsUnauthorised(e.Error)) {
+                SessionExpired();
+            }
+            else {
+                GetCDRsCountComplete(e);
+            }
         }
 
         public override void GetCDRsAsync(string whereExpression, int offset, int count) {
@@ -337,7 +445,28 @@ namespace SIPSorcery.Persistence
         }
 
         private void GetCDRsCompleted(object sender, GetCDRsCompletedEventArgs e) {
-            GetCDRsComplete(e);
+            if (IsUnauthorised(e.Error)) {
+                SessionExpired();
+            }
+            else {
+                GetCDRsComplete(e);
+            }
+        }
+
+        public override void CreateCustomerAsync(Customer customer) {
+            m_provisioningServiceProxy.CreateCustomerAsync(customer);
+        }
+
+        private void CreateCustomerCompleted(object sender, AsyncCompletedEventArgs e) {
+            CreateCustomerComplete(e);
+        }
+
+        public override void DeleteCustomerAsync(string customerUsername) {
+            m_provisioningServiceProxy.DeleteCustomerAsync(customerUsername);
+        }
+
+        private void DeleteCustomerCompleted(object sender, AsyncCompletedEventArgs e) {
+            DeleteCustomerComplete(e);
         }
     }
 }
